user signup

+++ controllers/projectController.js
@@ -10,6 +10,10 @@ const getProjects = asyncHandler(async(req,res) => {
 const addProject = asyncHandler(async(req,res) => {
     console.log(req.body);
     const {title , status , techStack, description, startDate, endDate, priority} = req.body;
+    if(!title || !status || !description){
+        res.status(400);
+        throw new Error("Fill necessary details");
+    }
     const project = await Project.create({
         title,
         status,
@@ -24,10 +28,19 @@ const addProject = asyncHandler(async(req,res) => {
 
 const getProject = asyncHandler(async(req,res) => {
     const project = await Project.findById(req.params.id);
+    if(!project){
+        res.status(404);
+        throw new Error("Project doesn't exists");
+    }
     res.status(200).json(project);
 });
 
 const updateProject = asyncHandler(async(req,res) => {
+    const project = await Project.findById(req.params.id);
+    if(!project){
+        res.status(404);
+        throw new Error("Project doesn't exists");
+    }
     const updatedProject = await Project.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -37,6 +50,11 @@ const updateProject = asyncHandler(async(req,res) => {
 });
 
 const deleteProject = asyncHandler(async(req,res) => {
+    const project = await Project.findById(req.params.id);
+    if(!project){
+        res.status(404);
+        throw new Error("Project doesn't exists");
+    }
     await Project.deleteOne({_id: req.params.id})
     res.status(202).json({message: `Deleted project with id: ${req.params.id}`});
 });
